Add deletion_pending and new_and_unlinked readonly attributes to the onbase_document record JSON.  Expose these new fields as facets on the search page (I18n this facet groups and facet terms nicely too).
If deletion_pending, show this in the display_string, a message on the readonly view and remove the "Delete" button from the toolbar.

If new_and_unlinked show a message on the readonly view and remove the "Delete" button from the toolbar.

Add a migration to reindex all onbase_documents to pick up the new display_string bits.